Renamed 'string_to_line_items' files to 'line_itemizer'
